+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- PrepareNextVersion: Specify published version from outside
+
 ## [3.0.11] - 2024-04-11
 
 ### Changed
@@ -135,6 +141,7 @@
 
 - Initial version.
 
+[Unreleased]: https://github.com/inlavigo/gg_publish/compare/3.0.11...HEAD
 [3.0.11]: https://github.com/inlavigo/gg_publish/compare/3.0.10...3.0.11
 [3.0.10]: https://github.com/inlavigo/gg_publish/compare/3.0.9...3.0.10
 [3.0.9]: https://github.com/inlavigo/gg_publish/compare/3.0.8...3.0.9